+++ platform/android/dnsproxy/lib/src/main/java/com/adguard/dnslibs/proxy/DnsProxyInitException.java
@@ -1,0 +1,14 @@
+package com.adguard.dnslibs.proxy;
+
+public class DnsProxyInitException extends RuntimeException {
+    private final DnsProxy.InitErrorCode code;
+
+    public DnsProxyInitException(DnsProxy.InitResult result) {
+        super(result.description);
+        this.code = result.code;
+    }
+
+    public DnsProxy.InitErrorCode getCode() {
+        return code;
+    }
+}